Refactor Boilerplate signalr implementation (#8557)
