+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Pages/Identity/Profile/ProfilePage.razor
@@ -39,7 +39,3 @@
         <DeleteAccountSection />
     </div>
 </div>
-
-@*#if (signalr == true)*@
-<BitSnackBar @ref="snackBar" />
-@*#endif*@
