+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Routes.razor
@@ -1,5 +1,9 @@
-﻿<CascadingAuthenticationState>
+﻿@*+:cnd:noEmit*@
+<CascadingAuthenticationState>
 	<AuthenticationStateLogger />
+    @*#if (signalr == true)*@
+    <SignalRHandler />
+    @*#endif*@
     <LayoutView Layout="@typeof(MainLayout)">
         <Router AppAssembly="@GetType().Assembly"
                 AdditionalAssemblies="@(AssemblyLoadContext.Default.Assemblies.Where(asm => asm.GetName().Name?.Contains("Boilerplate") is true))">
