+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Program.Middlewares.cs
@@ -65,7 +65,7 @@ public static partial class Program
         }).WithTags("Test");
 
         //#if (signalr == true)
-        app.MapHub<Hubs.IdentityHub>("/identity-hub");
+        app.MapHub<Hubs.AppHub>("/identity-hub");
         //#endif
 
         app.MapControllers().RequireAuthorization();
