+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Web/Program.Middlewares.cs
@@ -113,7 +113,7 @@ public static partial class Program
         }).WithTags("Test");
 
         //#if (signalr == true)
-        app.MapHub<Api.Hubs.IdentityHub>("/identity-hub");
+        app.MapHub<Api.Hubs.AppHub>("/identity-hub");
         //#endif
 
         app.MapControllers().RequireAuthorization();
