+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Shared/Resources/AppStrings.fr.resx
@@ -235,11 +235,8 @@
   </data>
     <!--#endif-->
     <!--#if (signalr == true) -->
-    <data name="NewUserSessionSnackbarTitle" xml:space="preserve">
-    <value>Nouvelle session !</value>
-  </data>
-    <data name="DeviceDetails" xml:space="preserve">
-    <value>Appareil : {0}</value>
+    <data name="TwoFactorTokenPushText" xml:space="preserve">
+    <value>{0} est votre jeton à deux facteurs dans Boilerplate.</value>
   </data>
     <!--#endif-->
     <data name="FullName" xml:space="preserve">
